IMDB Sentiment Analysis using RNN (LSTM)
This project applies a Recurrent Neural Network (RNN) using Long Short-Term Memory (LSTM) layers to perform sentiment analysis on the IMDB movie reviews dataset. The dataset is provided by Keras and consists of 50,000 movie reviews labeled as positive or negative. The goal is to build a deep learning model capable of classifying the sentiment of reviews.